ingestion/python/requirements.txt
PyGithub
kafka-python

